backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
SQLAlchemy==2.0.32
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
redis==5.0.8
PyJWT==2.10.0
alembic==1.14.0
inject==5.2.1
flake8==7.2.0
flake8-docstrings==1.7.0